// src/app/p-app/p-purchase/shared/dto/DTODomesticOrders.dto.ts
export class DTODomesticOrders {
    Code: number = 0
    SupplierName: string = ''
    OrderNo: string = ''
    SupplierID: string = ''
    POID: string = ''
    OrderedTime: Date | string
    CreatedTime: Date | string
    EffDate: Date | string
    RemarkSupplier: string = ''
    RemarkWarehouse: string = ''
    SKU: number = 0
    TotalQuantity: number = 0
    Amount: number = 0
    CreatedBy: string = ''
    EstDeliveredTime: Date | string
    FinishedTime: Date | string
    StatusID: number = 1
    StatusName: string = 'Đang soạn thảo'
    OrderTypeID: number = 1
    OrderTypeName: string = ''
    Supplier: number = 0
    AmountAfterVAT: number = 0
    VAT: number = 0
    AmountBeforeVAT: number = 0
    EmailTo: string = ''
    EmailCc: string = ''
    EmailContent: string = ''
    IsSendEmail: boolean = true
    POURLLink: string = ''
    EstDeliveredPeriod: string = ''
    ReceivingFromOrdered: number = 0
    ReceivingPeriod: number = 0
    ReceivingStart:Date | string
    ReceivingEnd: Date | string
    Contact: string = ''
    ContactPhone: string = ''
    MultiDelivery: boolean = false
    DiscountItemAmount: number = 0
    DiscountPOAmount: number = 0
    DiscountPOPercent: number = 0
    IsSendAuto: boolean = false
    DeliveryLocation: number = 0
    DeliveryAddress: string = ''
    Date: Date | string
    AcceptBeforeTime: Date | string
    RefusedTime: Date | string
    AcceptedTime: Date | string
    VATAmount: number = 0
    ConfirmedSKU: number = 0
    ConfirmedQuantity: number = 0
    ConfirmedBeforeVAT: number = 0
    ConfirmedDiscountItem: number = 0
    ConfirmedDiscountPO: number = 0
    ConfirmedVAT: number = 0
    ConfirmedAfterVAT: number = 0
    RefuseReasonID: number = 0
    RefuseReason: string = ''
    WHName: string = ''
    RefuseReasonName: string = ''
    CancelReasonName: string = ''
    AcceptedBeforeTime: string = ''
}